Fixed not to crash after deleting an entry.
In the previous version, even when an entry is deleted, the application
did not update the list in .photonote/groups/<group_name>. So if a user
does an operation which reads data again from .photonote (these
operations are deleting a group and adding a group) immediately after
deleting an entry, the application crashed.

This patch changes to write all groups after deleting an entry. This is
not the best. The best is updating only the group of the deleted entry.